requirements.txt
alembic==1.6.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
# Editable install with no version control (draft==0.0.0)
-e /Users/barry/src/Flask/flask-draft-sse/src
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
Flask-WTF==0.15.1
greenlet==1.0.0
gunicorn==20.1.0
idna==3.1
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
pip-review==1.1.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.19
sseclient==0.0.27
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
